feat: setup vector to export Kubernetes logs to Axiom
This is a first stab at exporting the logs to Kubernetes. It's mainly
focused on `control-api` and tries to parse the JSON logs, assigning the
fields correctly. We still export the message from the other services,
but since we don't export the traces and don't emit JSON, we can't have
the same experience as we have in `control-api`.

It limits the export of Kubes log to the default namespace, since we're
not managing the K8s cluster ourselves. The proper way to do it would
either be to use the `vector.dev/enabled` label (or something like that)
on each pod/deployment, or to define a custom Kubernetes label to define
whether or not we want to export the logs.

Everything is currently exported under the same dataset (and with the
same token). I'm not sure that's the right approach, because logs
exported from Kubes and traces don't have the same fields, but I suggest
we run with that for now in order to better understand how to organize
all of that.