example: add `fidget` package to the API route
Not sure this is how we wan't to support it, because it's going to
generate a big single Javascript file with all the dependencies, but
since we're using `esbuild` this is the easiest way to get dependencies
to work (i.e. doing nothing more than adding a package.json inside the
`api` folder). 

In the future, I think we might want to look into using Deno directly
instead of building the files. Or maybe we want to package the
dependencies on the side.. anyway, that's for future us.